+++ src/auth/passkey_authenticator.cpp
@@ -52,14 +52,14 @@ namespace {
 
 /// @brief Read the CBOR argument (length or integer payload) and advance pos.
 static size_t passkeyCborReadArg(const std::vector<uint8_t>& d, size_t pos, uint64_t& out) {
-    if (pos >= d.size()) {
+    if (pos >= static_cast<int>(d.size())) {
         throw std::runtime_error("CBOR: truncated data");
     }
     const uint8_t info = d[pos] & 0x1F;
     ++pos;
     if (info <= 23) { out = info; return pos; }
     if (info == 24) {
-        if (pos >= d.size()) {
+        if (pos >= static_cast<int>(d.size())) {
           throw std::runtime_error("CBOR: truncated 1-byte arg");
         }
         out = d[pos++]; return pos;
@@ -97,7 +97,7 @@ static size_t passkeyCborReadArg(const std::vector<uint8_t>& d, size_t pos, uint
 
 /// @brief Skip one CBOR item, returning the new position.
 static size_t passkeyCborSkip(const std::vector<uint8_t>& d, size_t pos) {
-    if (pos >= d.size()) {
+    if (pos >= static_cast<int>(d.size())) {
       throw std::runtime_error("CBOR: truncated (skip)");
     }
     const uint8_t initial = d[pos];
@@ -242,7 +242,7 @@ static void passkeyCborParseCoseKey(const std::vector<uint8_t>& d, size_t pos, P
     pos = passkeyCborReadArg(d, pos, count);
 
     for (uint64_t i = 0; i < count; ++i) {
-        if (pos >= d.size()) {
+        if (pos >= static_cast<int>(d.size())) {
           throw std::runtime_error("CBOR: truncated COSE key map");
         }
 
@@ -257,7 +257,7 @@ static void passkeyCborParseCoseKey(const std::vector<uint8_t>& d, size_t pos, P
             pos = passkeyCborSkip(d, pos); pos = passkeyCborSkip(d, pos); continue;
         }
 
-        if (pos >= d.size()) {
+        if (pos >= static_cast<int>(d.size())) {
           throw std::runtime_error("CBOR: truncated COSE key value");
         }
         const uint8_t v_major = d[pos] >> 5;
